requirements.txt
scikit-learn>=1.0
pandas>=1.3
numpy>=1.21
matplotlib>=3.5
plotly>=5.0
umap-learn>=0.5
jinja2>=3.0
